requirements.txt
bokeh
Pillow
flask
colour
gspread
matplotlib
numpy
scipy
oauth2client
tqdm
pygame
pyPDF2

autopep8
pylint
jedi
neovim
jupyter
